stripe_assignment/requirement.txt
Django==3.1.6
stripe==2.55.2
djangorestframework=3.13.1
python-dotenv=0.20.0